requirements.txt
websockets==8.1
colorama==0.4.4
python-valve==0.2.1